pathfinding/requirements.txt
psycopg2-binary
pandas
pygame
requests